Feat/config validator

+++ Colibri/Runtime/Pipeline/RoutingEngine/ParamValue.cs
@@ -2,10 +2,10 @@ namespace Colibri.Runtime.Pipeline.RoutingEngine;
 
 public readonly struct ParamValue
 {
-    public readonly ushort Start;
-    public readonly byte Length;
+    public readonly int Start;
+    public readonly ushort Length;
 
-    public ParamValue(ushort start, byte length)
+    public ParamValue(int start, ushort length)
     {
         Start = start;
         Length = length;
